visible setting changes

+++ src/ask.h
@@ -23,7 +23,7 @@
 #define err(x) Rf_errorcall(R_NilValue, x)
 #define warn(x) Rf_warningcall(R_NilValue, x)
 
-int get_threads(const int64_t n, const bool throttle);
+int get_threads(const int n, const bool throttle);
 void init_ask_threads(void);
 
 // API
